[DOCS] Update docs about generated licenses. (elastic/x-pack#5134)
This is related to elastic/x-pack-elasticsearch#1941. This updates the
x-pack documentation to describe the fact that a user can not generate
a trial or basic license at node startup.

Additionally, it updates the documenation to describe how to start a
trial.

Original commit: elastic/x-pack@b9ce9ac08f21033dcb861f463779c308ddae78c8