Setup CONSOLE testing
Sets up the Elasticsearch doc build testing plugin so they can be
automatically tested. You can do it with:
```
./gradlew check
```

Right now these use the oss distribution so the tests fail because the
tests rely on xpack. Installing xpack is a bit difficult so I propose we
wait until 6.3 is released and change the version to 6.3.