Update flightdata-classification.asciidoc
This paragraph is explaining the bottom row of a confusion matrix.  It correctly states the first column as 'false negative,' but then incorrectly says the second column (i.e. bottom right) is a true positive. TP is in the upper left.  See https://en.wikipedia.org/wiki/Confusion_matrix    It's not called a 'confusion' matrix for nothing!   Easy to mislabel. This is particularly confusing (haha) because the categories are True and False, which muddles the idea of Positive (row 1) vs Negative (row 2) predictions.