[DOCS] Updates guide to exportint/importing models
This version is a suggestion to make the instructions more concise and
to make use of only two tools (curl, jq) from the command line. This
makes prerequisites fewer, and following along easier. There are some
minor grammatical changes to the section on importing models trained
outside the Stack which hopefully also makes the terminology a bit more
in line with common parlance.