[DOCS] Update ML link for ES #68051
PR [#68051](https://github.com/elastic/elasticsearch/pull/68051) in the Elasticsearch repo is making changes to the networking docs, which impacts the HTTP settings. Specifically, `modules-http` is changing to `modules-network`, with the HTTP settings nested underneath. 

I checked out that PR locally and built the docs, and the resulting URL for the new HTTP section results in `.../guide/modules-network.html#http-settings`. This ML link must be updated for the ES build to pass.