[DOCS] Add notes about security for ML anomaly detection job results
ML anomaly detection jobs cause certain field values from the
source indices to be copied into the results indices when an
anomaly is detected.

This PR calls out the security implications of that. If an ML
anomaly detection job is being run against sensitive data then
it is important to carefully consider who has access to ML
anomaly detection results.