Update installing-stack-demo-self.asciidoc to mitigate error-handling issue. update variable on tail command (#3015)
* Update installing-stack-demo-self.asciidoc

Adding an Important clarification. The enrollment script fails without proper error handling if the second node is not able to access the first node. the elasticsearch.yml ends up modified and the enrollment script is unable to be reran without undoing all the changes. Reminder added to ensure network access is available before proceeding.

second. the tail command is incorrect unless the cluster.name is also elasticsearch-demo. this value defaults to cluster.name

* nit to trigger CI

---------

Co-authored-by: Liam Thompson <32779855+leemthompo@users.noreply.github.com>